requirements.txt
appdirs==1.4.4
asgiref==3.6.0
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
dj-database-url==1.3.0
Django==4.1.5
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.1
fix-yahoo-finance==0.1.37
frozendict==2.3.7
gunicorn==20.1.0
h11==0.14.0
html5lib==1.1
idna==3.4
install==1.3.5
lxml==4.9.2
multitasking==0.0.11
numpy==1.24.3
outcome==1.2.0
pandas==2.0.1
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.6.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.29.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.3
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
webencodings==0.5.1
whitenoise==6.4.0
wsproto==1.2.0
